refactor: rename services folder to api for better clarity
- Rename src/services/ to src/api/ to better reflect API communication purpose
- Update import path in admin-review-actions.ts
- Improves code organization and makes folder purpose more explicit
